\name{asia}
\title{asia}
\alias{asia} 

\description{Asia data.}
\usage{data(asia)}

\format{
     Asia data set
\describe{
  \item{AUT}{the dichotomous outcome of interest indicating whether not there was
     at least one violent terrorist act in a country/year pair.}
  \item{DEM}{the Polity IV 21-point democracy scale ranging from -10 indicating a hereditary monarchy to +10 indicating a fully consolidated democracy.  (Gurr, Marshall, and Jaggers 2003).}
  \item{FED}{is assigned zero if sub-national governments do not have substantial
     taxing, spending, and regulatory authority, and one otherwise.}
  \item{SYS}{coded as: (0) for direct presidential elections, (1) for strong
     president elected by assembly, and (2) dominant parliamentary
     government.}
  \item{AUT}{a dichotomous variable indicating whether or not there are
     autonomous regions not directly controlled by central government. }
}
}


\references{Koch, M. T. Cranmer, S.  (2007). "Terrorism than Governments of the Right?  Testing the 'Dick Cheney'
Hypothesis: Do Governments of the Left Attract More than Goverments of the
Right?'' \emph{Conflict Management and Peace Science} 24, 311-326.


  
}


\keyword{datasets}

