\name{ssas}
\title{ssas}
\alias{ssas}

\description{Scottish Social Attitudes Survey, 2006 (UK Data Archive Study Number 5840).}
\usage{data(ssas)}

\format{

\describe{
  \item{scotpar2}{the dichotomous outcome indicating whether not respondents favored full independence.}
  \item{hedqual2}{the respondent's education level.}
  \item{hincdif2}{the degree to which the respondent is living comfortably on current income or not.}  
  \item{househld}{the number of people living in the respondent's household.}
  \item{idlosem}{a dichotomous variable indicating whether or not the respondent agreed with the statement that increased numbers of Muslims in Scotland would erode the national identity.}
  \item{marrmus}{a dichotomous variable indicating whether or not the respondent would be unhappy or very unhappy if a family member married a Muslim.}
  \item{natinnat}{agreement that the Scottish Executive works in Scotland's long-term interests.}
  \item{nhssat}{a dichotomous variable indicating the respondent's satisfication (1) or dissatisfication (0) with the National Health Service.}
  \item{ptyallgs}{party allegiance with the ordering of parties given from more conservative to more liberal.}
  \item{rage}{the age of the respondent.}
  \item{relgsums}{a dichotomous variable indicating identification with the Church of Scotland versus another or no religion.}
  \item{rsex}{the gender of the respondent.}
  \item{ukintnat}{agreement that the UK government works in Scotland's long-term interests.}
  \item{unionsa}{a dichotomous variable indicating the respondent's union membership at work.}
  \item{voiceuk3}{the respondent's belief that the Scottish Parliament gives Scotland a greater voice in the UK.}
  \item{whrbrn}{a dichotomous variable indicating birth in Scotland or not.}

}
}



\keyword{datasets}

